requirements.txt
jsonschema>=2.6.0
tqdm>=4.43.0
opencv-python>=3.4.8.29
albumentations=0.4.5
imgaug=0.4.0
pytorch>=1.4.0
torchvision>=0.5.0
numpy<=1.17.4
